# Conflicts:
#	build.gradle

+++ treasury-impl/src/main/java/com/mislbd/ababil/treasury/repository/schema/TransactionRecordEntity.java
@@ -1,5 +1,6 @@
 package com.mislbd.ababil.treasury.repository.schema;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import javax.persistence.*;
 import lombok.AllArgsConstructor;
@@ -36,7 +37,7 @@ public class TransactionRecordEntity {
   private LocalDate txnDate;
 
   @Column(name = "TRAMOUNT")
-  private int amount;
+  private BigDecimal amount;
 
   @Column(name = "TRORIGINATINGBRID")
   private String trOiginatingBrId;
@@ -45,16 +46,16 @@ public class TransactionRecordEntity {
   private String instrumentNo;
 
   @Column(name = "TRTRDID")
-  private int txnDefId;
+  private Long txnDefId;
 
   @Column(name = "TRREFACCID")
-  private int refAccId;
+  private Long refAccId;
 
   @Column(name = "TRNARRATION")
   private String narration;
 
   @Column(name = "TRGLOBALTXNNO")
-  private int globalTxnNo;
+  private Long globalTxnNo;
 
   @Column(name = "TRINITIATOR")
   private String initiator;
